backend/requirements.txt
# FinanceCommander AI Portal Backend v2.0 Dependencies
# Aligned with root requirements.txt — keep in sync.

# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9
alembic>=1.13.3

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.22

# LLM Providers
openai~=1.83.0
anthropic>=0.79.0
google-generativeai>=0.4.0
litellm>=1.40.0

# Pipeline orchestration
crewai>=0.80.0

# WebSockets
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0
tenacity>=9.0.0

# Console Intelligence (SSH)
asyncssh>=2.14.0
pyyaml>=6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
